Made panels fit grid size and improvements to the GUI
