+++ src/graphics/panels/TetrisGraphics.java
@@ -1,5 +1,7 @@
 package graphics.panels;
 
+import java.awt.Color;
+
 import javax.swing.JPanel;
 
 import graphics.Frame;
@@ -12,6 +14,7 @@ public class TetrisGraphics extends JPanel {
 		this.frame = parent;
 		this.setLayout(null);
 		this.setBounds(0, 0, 900, 940);
+		this.setBackground(Color.BLACK);
 		JPanel titlePanel = new TitlePanel(this);
 		JPanel gamePanel = new GamePanel(this);
 		JPanel displayPanel = new DisplayPanel(this);
@@ -22,5 +25,5 @@ public class TetrisGraphics extends JPanel {
 
 	public Frame getFrame() {
 		return frame;
-	}	
+	}
 }
